audittools: move example code from package doc into a standalone example
Also, make it compile.
