+++ audittools/trail.go
@@ -39,8 +39,9 @@ type AuditTrail struct {
 
 // Commit takes a AuditTrail that receives audit events from an event sink and publishes them to
 // a specific RabbitMQ Connection using the specified amqp URI and queue name.
-// The OnSuccessfulPublish and OnFailedPublish closures are executed as per
-// their respective case.
+// The OnSuccessfulPublish and OnFailedPublish closures are executed as per their respective case.
+//
+// This function blocks the current goroutine forever. It should be invoked with the "go" keyword.
 func (t AuditTrail) Commit(ctx context.Context, rabbitmqURI url.URL, rabbitmqQueueName string) {
 	rc, err := newRabbitConnection(rabbitmqURI, rabbitmqQueueName)
 	if err != nil {
